Improve pty performance
Current implementation is using IO#getch. It's a wrong choise. Getch
constantly switches between raw and cooked mode, while having a timeout.
There is a small chance that the key press is received in raw mode, with
echo on. That creates a inconsistency between Pry thread's output via
PTY and the actual display on the screen. The issue opportunity
signifies when the keyboard repeat frequence is high. To solve this, I
force everything in raw mode, and switch to cooked mode when a line of
code is being evaluated.