Implement VCR-style integration test
The previous way of doing integration test is to copy, paste, and mark
actual screen into the test file. That process is time-consuming,
espeically when there are tiny changes on the screen (change arrow
symbol for example). Thinking back to the original purpose of
integration tests is that they are meant to confirm the look-and-feel of
the UI components and whole real user interfaces while using the
debugger in real life. The detailed edge-cases correctness are handled
by unit tests. The manual effort wasted for recording and comparation
should be mininal, so that we can focus on other more important steps.
Therefore, when implementing integration tests should be think -> write
scenarios -> record -> compare/check -> confirm loop. Creating a
recorder to record sent keys, screen contents to compare in future is a
good idea.